First pass of normalization complete, handling all non-alias normalization
