+++ fleece/decoders/aarch64_common.C
@@ -1,537 +1,202 @@
 
-#include "Alias.h"
+/*
+ * See fleece/COPYRIGHT for copyright information.
+ *
+ * This file is a part of Fleece.
+ *
+ * Fleece is free software; you can redistribute it and/or modify it under the
+ * terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 3.0 of the License, or (at your option)
+ * any later version.
+ *  
+ * This software is distributed in the hope that it will be useful, but WITHOUT 
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ * 
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this software; if not, see www.gnu.org/licenses
+*/
+
 #include "aarch64_common.h"
 
-void addSysregAliases() {
+void removeExtraZeroesFromFmovImm(char* buf, int bufLen) {
    
-   // We only want to execute this once, for speed purposes. It has no effect
-   // on the actual aliases to execute it multiple times.
-   static bool done = false;
-   if (done) {
+   // Verify that this is an fmov instruction.
+   if (strncmp(buf, "fmov", 4)) {
       return;
    }
-   done = true;
-   Alias::addAlias("trcprgctlr", "s2_1_c0_c1_0");
-   Alias::addAlias("trcprocselr", "s2_1_c0_c2_0");
-   Alias::addAlias("trcstatr", "s2_1_c0_c3_0");
-   Alias::addAlias("trcconfigr", "s2_1_c0_c4_0");
-   Alias::addAlias("trcauxctlr", "s2_1_c0_c6_0");
-   Alias::addAlias("trceventctl0r", "s2_1_c0_c8_0");
-   Alias::addAlias("trcstallctlr", "s2_1_c0_c11_0");
-   Alias::addAlias("trctsctlr", "s2_1_c0_c12_0");
-   Alias::addAlias("trcsyncpr", "s2_1_c0_c13_0");
-   Alias::addAlias("trcccctlr", "s2_1_c0_c14_0");
-   Alias::addAlias("trcbbctlr", "s2_1_c0_c15_0");
-   Alias::addAlias("s3_5_c1_c0_0", "sctlr_el12");
-   Alias::addAlias("trcrsctlr2", "s2_1_c1_c2_0");
-   Alias::addAlias("trcacvr0", "s2_1_c2_c0_0");
-   Alias::addAlias("s3_5_c2_c0_0", "ttbr0_el12");
-   Alias::addAlias("trccidcvr0", "s2_1_c3_c0_0");
-   Alias::addAlias("s3_5_c4_c0_0", "spsr_el12");
-   Alias::addAlias("icc_pmr_el1", "s3_0_c4_c6_0");
-   Alias::addAlias("s3_5_c5_c1_0", "afsr0_el12");
-   Alias::addAlias("s3_5_c5_c2_0", "esr_el12");
-   Alias::addAlias("s3_0_c5_c3_0", "erridr_el1");
-   Alias::addAlias("s3_0_c5_c4_0", "erxfr_el1");
-   Alias::addAlias("s3_0_c5_c5_0", "erxmisc0_el1");
-   Alias::addAlias("s3_5_c6_c0_0", "far_el12");
-   Alias::addAlias("s3_0_c9_c9_0", "pmscr_el1");
-   Alias::addAlias("s3_0_c9_c10_0", "pmblimitr_el1");
-   Alias::addAlias("s3_5_c10_c2_0", "mair_el12");
-   Alias::addAlias("s3_5_c10_c3_0", "amair_el12");
-   Alias::addAlias("s3_5_c12_c0_0", "vbar_el12");
-   Alias::addAlias("icc_iar0_el1", "s3_0_c12_c8_0");
-   Alias::addAlias("ich_ap0r0_el2", "s3_4_c12_c8_0");
-   Alias::addAlias("icc_ap1r0_el1", "s3_0_c12_c9_0");
-   Alias::addAlias("ich_hcr_el2", "s3_4_c12_c11_0");
-   Alias::addAlias("ich_lr0_el2", "s3_4_c12_c12_0");
-   Alias::addAlias("icc_seien_el1", "s3_0_c12_c13_0");
-   Alias::addAlias("s3_5_c14_c1_0", "cntkctl_el12");
-   Alias::addAlias("s3_5_c14_c2_0", "cntp_tval_el02");
-   Alias::addAlias("s3_4_c14_c3_0", "cnthv_tval_el2");
-   Alias::addAlias("s3_5_c14_c3_0", "cntv_tval_el02");
-   Alias::addAlias("trcprgctlr", "s2_1_c0_c1_0");
-   Alias::addAlias("trcprocselr", "s2_1_c0_c2_0");
-   Alias::addAlias("trcstatr", "s2_1_c0_c3_0");
-   Alias::addAlias("trcconfigr", "s2_1_c0_c4_0");
-   Alias::addAlias("trcauxctlr", "s2_1_c0_c6_0");
-   Alias::addAlias("trceventctl0r", "s2_1_c0_c8_0");
-   Alias::addAlias("trcstallctlr", "s2_1_c0_c11_0");
-   Alias::addAlias("trctsctlr", "s2_1_c0_c12_0");
-   Alias::addAlias("trcsyncpr", "s2_1_c0_c13_0");
-   Alias::addAlias("trcccctlr", "s2_1_c0_c14_0");
-   Alias::addAlias("trcbbctlr", "s2_1_c0_c15_0");
-   Alias::addAlias("s3_5_c1_c0_0", "sctlr_el12");
-   Alias::addAlias("trcrsctlr2", "s2_1_c1_c2_0");
-   Alias::addAlias("trcacvr0", "s2_1_c2_c0_0");
-   Alias::addAlias("s3_5_c2_c0_0", "ttbr0_el12");
-   Alias::addAlias("trccidcvr0", "s2_1_c3_c0_0");
-   Alias::addAlias("s3_5_c4_c0_0", "spsr_el12");
-   Alias::addAlias("icc_pmr_el1", "s3_0_c4_c6_0");
-   Alias::addAlias("s3_5_c5_c1_0", "afsr0_el12");
-   Alias::addAlias("s3_5_c5_c2_0", "esr_el12");
-   Alias::addAlias("s3_0_c5_c3_0", "erridr_el1");
-   Alias::addAlias("s3_0_c5_c4_0", "erxfr_el1");
-   Alias::addAlias("s3_0_c5_c5_0", "erxmisc0_el1");
-   Alias::addAlias("s3_5_c6_c0_0", "far_el12");
-   Alias::addAlias("s3_0_c9_c9_0", "pmscr_el1");
-   Alias::addAlias("s3_0_c9_c10_0", "pmblimitr_el1");
-   Alias::addAlias("s3_5_c10_c2_0", "mair_el12");
-   Alias::addAlias("s3_5_c10_c3_0", "amair_el12");
-   Alias::addAlias("s3_5_c12_c0_0", "vbar_el12");
-   Alias::addAlias("icc_iar0_el1", "s3_0_c12_c8_0");
-   Alias::addAlias("ich_ap0r0_el2", "s3_4_c12_c8_0");
-   Alias::addAlias("icc_ap1r0_el1", "s3_0_c12_c9_0");
-   Alias::addAlias("ich_hcr_el2", "s3_4_c12_c11_0");
-   Alias::addAlias("ich_lr0_el2", "s3_4_c12_c12_0");
-   Alias::addAlias("icc_seien_el1", "s3_0_c12_c13_0");
-   Alias::addAlias("s3_5_c14_c1_0", "cntkctl_el12");
-   Alias::addAlias("s3_5_c14_c2_0", "cntp_tval_el02");
-   Alias::addAlias("s3_4_c14_c3_0", "cnthv_tval_el2");
-   Alias::addAlias("s3_5_c14_c3_0", "cntv_tval_el02");
-   Alias::addAlias("trctraceidr", "s2_1_c0_c0_1");
-   Alias::addAlias("trcqctlr", "s2_1_c0_c1_1");
-   Alias::addAlias("trcvmidcvr0", "s2_1_c3_c0_1");
-   Alias::addAlias("s3_5_c4_c0_1", "elr_el12");
-   Alias::addAlias("s3_0_c5_c3_1", "errselr_el1");
-   Alias::addAlias("s3_0_c5_c4_1", "erxctlr_el1");
-   Alias::addAlias("s3_0_c9_c10_1", "pmbptr_el1");
-   Alias::addAlias("s3_0_c12_c1_1", "disr_el1");
-   Alias::addAlias("s3_4_c12_c1_1", "vdisr_el2");
-   Alias::addAlias("ich_vtr_el2", "s3_4_c12_c11_1");
-   Alias::addAlias("s3_4_c13_c0_1", "contextidr_el2");
-   Alias::addAlias("s3_5_c14_c2_1", "cntp_ctl_el02");
-   Alias::addAlias("s3_4_c14_c3_1", "cnthv_ctl_el2");
-   Alias::addAlias("s3_5_c14_c3_1", "cntv_ctl_el02");
-   Alias::addAlias("trctraceidr", "s2_1_c0_c0_1");
-   Alias::addAlias("trcqctlr", "s2_1_c0_c1_1");
-   Alias::addAlias("trcvmidcvr0", "s2_1_c3_c0_1");
-   Alias::addAlias("s3_5_c4_c0_1", "elr_el12");
-   Alias::addAlias("s3_0_c5_c3_1", "errselr_el1");
-   Alias::addAlias("s3_0_c5_c4_1", "erxctlr_el1");
-   Alias::addAlias("s3_0_c9_c10_1", "pmbptr_el1");
-   Alias::addAlias("s3_0_c12_c1_1", "disr_el1");
-   Alias::addAlias("s3_4_c12_c1_1", "vdisr_el2");
-   Alias::addAlias("ich_vtr_el2", "s3_4_c12_c11_1");
-   Alias::addAlias("s3_4_c13_c0_1", "contextidr_el2");
-   Alias::addAlias("s3_5_c14_c2_1", "cntp_ctl_el02");
-   Alias::addAlias("s3_4_c14_c3_1", "cnthv_ctl_el2");
-   Alias::addAlias("s3_5_c14_c3_1", "cntv_ctl_el02");
-   Alias::addAlias("trcvictlr", "s2_1_c0_c0_2");
-   Alias::addAlias("trcviiectlr", "s2_1_c0_c1_2");
-   Alias::addAlias("trcvissctlr", "s2_1_c0_c2_2");
-   Alias::addAlias("trcvipcssctlr", "s2_1_c0_c3_2");
-   Alias::addAlias("s3_0_c0_c7_2", "id_aa64mmfr2_el1");
-   Alias::addAlias("trcvdctlr", "s2_1_c0_c8_2");
-   Alias::addAlias("trcvdsacctlr", "s2_1_c0_c9_2");
-   Alias::addAlias("trcvdarcctlr", "s2_1_c0_c10_2");
-   Alias::addAlias("trcssccr0", "s2_1_c1_c0_2");
-   Alias::addAlias("s3_5_c1_c0_2", "cpacr_el12");
-   Alias::addAlias("trcsscsr0", "s2_1_c1_c8_2");
-   Alias::addAlias("trcacatr0", "s2_1_c2_c0_2");
-   Alias::addAlias("s3_5_c2_c0_2", "tcr_el12");
-   Alias::addAlias("trccidcctlr0", "s2_1_c3_c0_2");
-   Alias::addAlias("trcvmidcctlr0", "s2_1_c3_c2_2");
-   Alias::addAlias("s3_0_c5_c4_2", "erxstatus_el1");
-   Alias::addAlias("s3_0_c9_c9_2", "pmsicr_el1");
-   Alias::addAlias("icc_hppir0_el1", "s3_0_c12_c8_2");
-   Alias::addAlias("ich_misr_el2", "s3_4_c12_c11_2");
-   Alias::addAlias("s3_5_c14_c2_2", "cntp_cval_el02");
-   Alias::addAlias("s3_4_c14_c3_2", "cnthv_cval_el2");
-   Alias::addAlias("s3_5_c14_c3_2", "cntv_cval_el02");
-   Alias::addAlias("trcvictlr", "s2_1_c0_c0_2");
-   Alias::addAlias("trcviiectlr", "s2_1_c0_c1_2");
-   Alias::addAlias("trcvissctlr", "s2_1_c0_c2_2");
-   Alias::addAlias("trcvipcssctlr", "s2_1_c0_c3_2");
-   Alias::addAlias("s3_0_c0_c7_2", "id_aa64mmfr2_el1");
-   Alias::addAlias("trcvdctlr", "s2_1_c0_c8_2");
-   Alias::addAlias("trcvdsacctlr", "s2_1_c0_c9_2");
-   Alias::addAlias("trcvdarcctlr", "s2_1_c0_c10_2");
-   Alias::addAlias("trcssccr0", "s2_1_c1_c0_2");
-   Alias::addAlias("s3_5_c1_c0_2", "cpacr_el12");
-   Alias::addAlias("trcsscsr0", "s2_1_c1_c8_2");
-   Alias::addAlias("trcacatr0", "s2_1_c2_c0_2");
-   Alias::addAlias("s3_5_c2_c0_2", "tcr_el12");
-   Alias::addAlias("trccidcctlr0", "s2_1_c3_c0_2");
-   Alias::addAlias("trcvmidcctlr0", "s2_1_c3_c2_2");
-   Alias::addAlias("s3_0_c5_c4_2", "erxstatus_el1");
-   Alias::addAlias("s3_0_c9_c9_2", "pmsicr_el1");
-   Alias::addAlias("icc_hppir0_el1", "s3_0_c12_c8_2");
-   Alias::addAlias("ich_misr_el2", "s3_4_c12_c11_2");
-   Alias::addAlias("s3_5_c14_c2_2", "cntp_cval_el02");
-   Alias::addAlias("s3_4_c14_c3_2", "cnthv_cval_el2");
-   Alias::addAlias("s3_5_c14_c3_2", "cntv_cval_el02");
-   Alias::addAlias("trcsspcicr0", "s2_1_c1_c0_3");
-   Alias::addAlias("s3_0_c4_c2_3", "pan");
-   Alias::addAlias("s3_4_c5_c2_3", "vsesr_el2");
-   Alias::addAlias("s3_0_c5_c4_3", "erxaddr_el1");
-   Alias::addAlias("s3_0_c9_c9_3", "pmsirr_el1");
-   Alias::addAlias("s3_0_c9_c10_3", "pmbsr_el1");
-   Alias::addAlias("icc_bpr0_el1", "s3_0_c12_c8_3");
-   Alias::addAlias("icc_rpr_el1", "s3_0_c12_c11_3");
-   Alias::addAlias("ich_eisr_el2", "s3_4_c12_c11_3");
-   Alias::addAlias("trcsspcicr0", "s2_1_c1_c0_3");
-   Alias::addAlias("s3_0_c4_c2_3", "pan");
-   Alias::addAlias("s3_4_c5_c2_3", "vsesr_el2");
-   Alias::addAlias("s3_0_c5_c4_3", "erxaddr_el1");
-   Alias::addAlias("s3_0_c9_c9_3", "pmsirr_el1");
-   Alias::addAlias("s3_0_c9_c10_3", "pmbsr_el1");
-   Alias::addAlias("icc_bpr0_el1", "s3_0_c12_c8_3");
-   Alias::addAlias("icc_rpr_el1", "s3_0_c12_c11_3");
-   Alias::addAlias("ich_eisr_el2", "s3_4_c12_c11_3");
-   Alias::addAlias("trcseqevr0", "s2_1_c0_c0_4");
-   Alias::addAlias("trcseqrstevr", "s2_1_c0_c6_4");
-   Alias::addAlias("trcseqstr", "s2_1_c0_c7_4");
-   Alias::addAlias("trcextinselr", "s2_1_c0_c8_4");
-   Alias::addAlias("s2_0_c1_c0_4", "oslar_el1");
-   Alias::addAlias("trcoslsr", "s2_1_c1_c1_4");
-   Alias::addAlias("trcpdcr", "s2_1_c1_c4_4");
-   Alias::addAlias("trcpdsr", "s2_1_c1_c5_4");
-   Alias::addAlias("trcdvcvr0", "s2_1_c2_c0_4");
-   Alias::addAlias("s3_0_c4_c2_4", "uao");
-   Alias::addAlias("trcitctrl", "s2_1_c7_c0_4");
-   Alias::addAlias("s3_0_c9_c9_4", "pmsfcr_el1");
-   Alias::addAlias("s3_3_c9_c12_4", "pmswinc_el0");
-   Alias::addAlias("ich_vseir_el2", "s3_4_c12_c9_4");
-   Alias::addAlias("icc_ctlr_el1", "s3_0_c12_c12_4");
-   Alias::addAlias("trcseqevr0", "s2_1_c0_c0_4");
-   Alias::addAlias("trcseqrstevr", "s2_1_c0_c6_4");
-   Alias::addAlias("trcseqstr", "s2_1_c0_c7_4");
-   Alias::addAlias("trcextinselr", "s2_1_c0_c8_4");
-   Alias::addAlias("s2_0_c1_c0_4", "oslar_el1");
-   Alias::addAlias("trcoslsr", "s2_1_c1_c1_4");
-   Alias::addAlias("trcpdcr", "s2_1_c1_c4_4");
-   Alias::addAlias("trcpdsr", "s2_1_c1_c5_4");
-   Alias::addAlias("trcdvcvr0", "s2_1_c2_c0_4");
-   Alias::addAlias("s3_0_c4_c2_4", "uao");
-   Alias::addAlias("trcitctrl", "s2_1_c7_c0_4");
-   Alias::addAlias("s3_0_c9_c9_4", "pmsfcr_el1");
-   Alias::addAlias("s3_3_c9_c12_4", "pmswinc_el0");
-   Alias::addAlias("ich_vseir_el2", "s3_4_c12_c9_4");
-   Alias::addAlias("icc_ctlr_el1", "s3_0_c12_c12_4");
-   Alias::addAlias("trccntrldvr0", "s2_1_c0_c0_5");
-   Alias::addAlias("trccntctlr0", "s2_1_c0_c4_5");
-   Alias::addAlias("trccntvr0", "s2_1_c0_c8_5");
-   Alias::addAlias("s3_0_c9_c9_5", "pmsevfr_el1");
-   Alias::addAlias("icc_sre_el2", "s3_4_c12_c9_5");
-   Alias::addAlias("ich_elsr_el2", "s3_4_c12_c11_5");
-   Alias::addAlias("trccntrldvr0", "s2_1_c0_c0_5");
-   Alias::addAlias("trccntctlr0", "s2_1_c0_c4_5");
-   Alias::addAlias("trccntvr0", "s2_1_c0_c8_5");
-   Alias::addAlias("s3_0_c9_c9_5", "pmsevfr_el1");
-   Alias::addAlias("icc_sre_el2", "s3_4_c12_c9_5");
-   Alias::addAlias("ich_elsr_el2", "s3_4_c12_c11_5");
-   Alias::addAlias("trcidr8", "s2_1_c0_c0_6");
-   Alias::addAlias("trcdvcmr0", "s2_1_c2_c0_6");
-   Alias::addAlias("trcclaimset", "s2_1_c7_c8_6");
-   Alias::addAlias("trcclaimclr", "s2_1_c7_c9_6");
-   Alias::addAlias("trcdevaff0", "s2_1_c7_c10_6");
-   Alias::addAlias("trclsr", "s2_1_c7_c13_6");
-   Alias::addAlias("trcauthstatus", "s2_1_c7_c14_6");
-   Alias::addAlias("trcdevarch", "s2_1_c7_c15_6");
-   Alias::addAlias("s3_0_c9_c9_6", "pmslatfr_el1");
-   Alias::addAlias("icc_igrpen0_el1", "s3_0_c12_c12_6");
-   Alias::addAlias("trcidr8", "s2_1_c0_c0_6");
-   Alias::addAlias("trcdvcmr0", "s2_1_c2_c0_6");
-   Alias::addAlias("trcclaimset", "s2_1_c7_c8_6");
-   Alias::addAlias("trcclaimclr", "s2_1_c7_c9_6");
-   Alias::addAlias("trcdevaff0", "s2_1_c7_c10_6");
-   Alias::addAlias("trclsr", "s2_1_c7_c13_6");
-   Alias::addAlias("trcauthstatus", "s2_1_c7_c14_6");
-   Alias::addAlias("trcdevarch", "s2_1_c7_c15_6");
-   Alias::addAlias("s3_0_c9_c9_6", "pmslatfr_el1");
-   Alias::addAlias("icc_igrpen0_el1", "s3_0_c12_c12_6");
-   Alias::addAlias("trcimspec0", "s2_1_c0_c0_7");
-   Alias::addAlias("trcdevid", "s2_1_c7_c2_7");
-   Alias::addAlias("trcdevtype", "s2_1_c7_c3_7");
-   Alias::addAlias("trcpidr4", "s2_1_c7_c4_7");
-   Alias::addAlias("trccidr0", "s2_1_c7_c12_7");
-   Alias::addAlias("s3_0_c9_c9_7", "pmsidr_el1");
-   Alias::addAlias("s3_0_c9_c10_7", "pmbidr_el1");
-   Alias::addAlias("ich_vmcr_el2", "s3_4_c12_c11_7");
-   Alias::addAlias("trcimspec0", "s2_1_c0_c0_7");
-   Alias::addAlias("trcdevid", "s2_1_c7_c2_7");
-   Alias::addAlias("trcdevtype", "s2_1_c7_c3_7");
-   Alias::addAlias("trcpidr4", "s2_1_c7_c4_7");
-   Alias::addAlias("trccidr0", "s2_1_c7_c12_7");
-   Alias::addAlias("s3_0_c9_c9_7", "pmsidr_el1");
-   Alias::addAlias("s3_0_c9_c10_7", "pmbidr_el1");
-   Alias::addAlias("ich_vmcr_el2", "s3_4_c12_c11_7");
-   Alias::addAlias("s3_0_c0_c0_0", "midr_el1");
-   Alias::addAlias("s3_1_c0_c0_0", "ccsidr_el1");
-   Alias::addAlias("trcprgctlr", "s2_1_c0_c1_0");
-   Alias::addAlias("s2_3_c0_c1_0", "mdccsr_el0");
-   Alias::addAlias("s3_0_c0_c1_0", "id_pfr0_el1");
-   Alias::addAlias("trcprocselr", "s2_1_c0_c2_0");
-   Alias::addAlias("s3_0_c0_c2_0", "id_isar0_el1");
-   Alias::addAlias("s3_0_c0_c3_0", "mvfr0_el1");
-   Alias::addAlias("trcconfigr", "s2_1_c0_c4_0");
-   Alias::addAlias("s3_0_c0_c4_0", "id_aa64pfr0_el1");
-   Alias::addAlias("dbgdtrtx_el0", "dbgdtrrx_el0");
-   Alias::addAlias("s3_0_c0_c5_0", "id_aa64dfr0_el1");
-   Alias::addAlias("trcauxctlr", "s2_1_c0_c6_0");
-   Alias::addAlias("s3_0_c0_c6_0", "id_aa64isar0_el1");
-   Alias::addAlias("s3_0_c0_c7_0", "id_aa64mmfr0_el1");
-   Alias::addAlias("trceventctl0r", "s2_1_c0_c8_0");
-   Alias::addAlias("trcstallctlr", "s2_1_c0_c11_0");
-   Alias::addAlias("trctsctlr", "s2_1_c0_c12_0");
-   Alias::addAlias("trcsyncpr", "s2_1_c0_c13_0");
-   Alias::addAlias("trcccctlr", "s2_1_c0_c14_0");
-   Alias::addAlias("trcbbctlr", "s2_1_c0_c15_0");
-   Alias::addAlias("s2_0_c1_c0_0", "mdrar_el1");
-   Alias::addAlias("s3_5_c1_c0_0", "sctlr_el12");
-   Alias::addAlias("trcrsctlr2", "s2_1_c1_c2_0");
-   Alias::addAlias("trcacvr0", "s2_1_c2_c0_0");
-   Alias::addAlias("s3_5_c2_c0_0", "ttbr0_el12");
-   Alias::addAlias("trccidcvr0", "s2_1_c3_c0_0");
-   Alias::addAlias("s3_5_c4_c0_0", "spsr_el12");
-   Alias::addAlias("icc_pmr_el1", "s3_0_c4_c6_0");
-   Alias::addAlias("s3_5_c5_c1_0", "afsr0_el12");
-   Alias::addAlias("s3_5_c5_c2_0", "esr_el12");
-   Alias::addAlias("s3_0_c5_c3_0", "erridr_el1");
-   Alias::addAlias("s3_0_c5_c4_0", "erxfr_el1");
-   Alias::addAlias("s3_0_c5_c5_0", "erxmisc0_el1");
-   Alias::addAlias("s3_5_c6_c0_0", "far_el12");
-   Alias::addAlias("s3_0_c9_c9_0", "pmscr_el1");
-   Alias::addAlias("s3_0_c9_c10_0", "pmblimitr_el1");
-   Alias::addAlias("s3_5_c10_c2_0", "mair_el12");
-   Alias::addAlias("s3_5_c10_c3_0", "amair_el12");
-   Alias::addAlias("s3_5_c12_c0_0", "vbar_el12");
-   Alias::addAlias("s3_0_c12_c1_0", "isr_el1");
-   Alias::addAlias("ich_ap0r0_el2", "s3_4_c12_c8_0");
-   Alias::addAlias("icc_ap1r0_el1", "s3_0_c12_c9_0");
-   Alias::addAlias("ich_hcr_el2", "s3_4_c12_c11_0");
-   Alias::addAlias("ich_lr0_el2", "s3_4_c12_c12_0");
-   Alias::addAlias("icc_seien_el1", "s3_0_c12_c13_0");
-   Alias::addAlias("s3_5_c14_c1_0", "cntkctl_el12");
-   Alias::addAlias("s3_5_c14_c2_0", "cntp_tval_el02");
-   Alias::addAlias("s3_4_c14_c3_0", "cnthv_tval_el2");
-   Alias::addAlias("s3_5_c14_c3_0", "cntv_tval_el02");
-   Alias::addAlias("s3_0_c0_c0_0", "midr_el1");
-   Alias::addAlias("s3_1_c0_c0_0", "ccsidr_el1");
-   Alias::addAlias("trcprgctlr", "s2_1_c0_c1_0");
-   Alias::addAlias("s2_3_c0_c1_0", "mdccsr_el0");
-   Alias::addAlias("s3_0_c0_c1_0", "id_pfr0_el1");
-   Alias::addAlias("trcprocselr", "s2_1_c0_c2_0");
-   Alias::addAlias("s3_0_c0_c2_0", "id_isar0_el1");
-   Alias::addAlias("s3_0_c0_c3_0", "mvfr0_el1");
-   Alias::addAlias("trcconfigr", "s2_1_c0_c4_0");
-   Alias::addAlias("s3_0_c0_c4_0", "id_aa64pfr0_el1");
-   Alias::addAlias("dbgdtrtx_el0", "dbgdtrrx_el0");
-   Alias::addAlias("s3_0_c0_c5_0", "id_aa64dfr0_el1");
-   Alias::addAlias("trcauxctlr", "s2_1_c0_c6_0");
-   Alias::addAlias("s3_0_c0_c6_0", "id_aa64isar0_el1");
-   Alias::addAlias("s3_0_c0_c7_0", "id_aa64mmfr0_el1");
-   Alias::addAlias("trceventctl0r", "s2_1_c0_c8_0");
-   Alias::addAlias("trcstallctlr", "s2_1_c0_c11_0");
-   Alias::addAlias("trctsctlr", "s2_1_c0_c12_0");
-   Alias::addAlias("trcsyncpr", "s2_1_c0_c13_0");
-   Alias::addAlias("trcccctlr", "s2_1_c0_c14_0");
-   Alias::addAlias("trcbbctlr", "s2_1_c0_c15_0");
-   Alias::addAlias("s2_0_c1_c0_0", "mdrar_el1");
-   Alias::addAlias("s3_5_c1_c0_0", "sctlr_el12");
-   Alias::addAlias("trcrsctlr2", "s2_1_c1_c2_0");
-   Alias::addAlias("trcacvr0", "s2_1_c2_c0_0");
-   Alias::addAlias("s3_5_c2_c0_0", "ttbr0_el12");
-   Alias::addAlias("trccidcvr0", "s2_1_c3_c0_0");
-   Alias::addAlias("s3_5_c4_c0_0", "spsr_el12");
-   Alias::addAlias("icc_pmr_el1", "s3_0_c4_c6_0");
-   Alias::addAlias("s3_5_c5_c1_0", "afsr0_el12");
-   Alias::addAlias("s3_5_c5_c2_0", "esr_el12");
-   Alias::addAlias("s3_0_c5_c3_0", "erridr_el1");
-   Alias::addAlias("s3_0_c5_c4_0", "erxfr_el1");
-   Alias::addAlias("s3_0_c5_c5_0", "erxmisc0_el1");
-   Alias::addAlias("s3_5_c6_c0_0", "far_el12");
-   Alias::addAlias("s3_0_c9_c9_0", "pmscr_el1");
-   Alias::addAlias("s3_0_c9_c10_0", "pmblimitr_el1");
-   Alias::addAlias("s3_5_c10_c2_0", "mair_el12");
-   Alias::addAlias("s3_5_c10_c3_0", "amair_el12");
-   Alias::addAlias("s3_5_c12_c0_0", "vbar_el12");
-   Alias::addAlias("s3_0_c12_c1_0", "isr_el1");
-   Alias::addAlias("ich_ap0r0_el2", "s3_4_c12_c8_0");
-   Alias::addAlias("icc_ap1r0_el1", "s3_0_c12_c9_0");
-   Alias::addAlias("ich_hcr_el2", "s3_4_c12_c11_0");
-   Alias::addAlias("ich_lr0_el2", "s3_4_c12_c12_0");
-   Alias::addAlias("icc_seien_el1", "s3_0_c12_c13_0");
-   Alias::addAlias("s3_5_c14_c1_0", "cntkctl_el12");
-   Alias::addAlias("s3_5_c14_c2_0", "cntp_tval_el02");
-   Alias::addAlias("s3_4_c14_c3_0", "cnthv_tval_el2");
-   Alias::addAlias("s3_5_c14_c3_0", "cntv_tval_el02");
-   Alias::addAlias("trctraceidr", "s2_1_c0_c0_1");
-   Alias::addAlias("s3_1_c0_c0_1", "clidr_el1");
-   Alias::addAlias("s3_3_c0_c0_1", "ctr_el0");
-   Alias::addAlias("trcqctlr", "s2_1_c0_c1_1");
-   Alias::addAlias("trcvmidcvr0", "s2_1_c3_c0_1");
-   Alias::addAlias("s3_5_c4_c0_1", "elr_el12");
-   Alias::addAlias("s3_0_c5_c3_1", "errselr_el1");
-   Alias::addAlias("s3_0_c5_c4_1", "erxctlr_el1");
-   Alias::addAlias("s3_0_c9_c10_1", "pmbptr_el1");
-   Alias::addAlias("s3_0_c12_c0_1", "rvbar_el1");
-   Alias::addAlias("s3_0_c12_c1_1", "disr_el1");
-   Alias::addAlias("s3_4_c12_c1_1", "vdisr_el2");
-   Alias::addAlias("icc_eoir0_el1", "s3_0_c12_c8_1");
-   Alias::addAlias("icc_dir_el1", "s3_0_c12_c11_1");
-   Alias::addAlias("s3_4_c13_c0_1", "contextidr_el2");
-   Alias::addAlias("s3_3_c14_c0_1", "cntpct_el0");
-   Alias::addAlias("s3_5_c14_c2_1", "cntp_ctl_el02");
-   Alias::addAlias("s3_4_c14_c3_1", "cnthv_ctl_el2");
-   Alias::addAlias("s3_5_c14_c3_1", "cntv_ctl_el02");
-   Alias::addAlias("trctraceidr", "s2_1_c0_c0_1");
-   Alias::addAlias("s3_1_c0_c0_1", "clidr_el1");
-   Alias::addAlias("s3_3_c0_c0_1", "ctr_el0");
-   Alias::addAlias("trcqctlr", "s2_1_c0_c1_1");
-   Alias::addAlias("trcvmidcvr0", "s2_1_c3_c0_1");
-   Alias::addAlias("s3_5_c4_c0_1", "elr_el12");
-   Alias::addAlias("s3_0_c5_c3_1", "errselr_el1");
-   Alias::addAlias("s3_0_c5_c4_1", "erxctlr_el1");
-   Alias::addAlias("s3_0_c9_c10_1", "pmbptr_el1");
-   Alias::addAlias("s3_0_c12_c0_1", "rvbar_el1");
-   Alias::addAlias("s3_0_c12_c1_1", "disr_el1");
-   Alias::addAlias("s3_4_c12_c1_1", "vdisr_el2");
-   Alias::addAlias("icc_eoir0_el1", "s3_0_c12_c8_1");
-   Alias::addAlias("icc_dir_el1", "s3_0_c12_c11_1");
-   Alias::addAlias("s3_4_c13_c0_1", "contextidr_el2");
-   Alias::addAlias("s3_3_c14_c0_1", "cntpct_el0");
-   Alias::addAlias("s3_5_c14_c2_1", "cntp_ctl_el02");
-   Alias::addAlias("s3_4_c14_c3_1", "cnthv_ctl_el2");
-   Alias::addAlias("s3_5_c14_c3_1", "cntv_ctl_el02");
-   Alias::addAlias("trcvictlr", "s2_1_c0_c0_2");
-   Alias::addAlias("trcviiectlr", "s2_1_c0_c1_2");
-   Alias::addAlias("s3_0_c0_c1_2", "id_dfr0_el1");
-   Alias::addAlias("trcvissctlr", "s2_1_c0_c2_2");
-   Alias::addAlias("trcvipcssctlr", "s2_1_c0_c3_2");
-   Alias::addAlias("trcvdctlr", "s2_1_c0_c8_2");
-   Alias::addAlias("trcvdsacctlr", "s2_1_c0_c9_2");
-   Alias::addAlias("trcvdarcctlr", "s2_1_c0_c10_2");
-   Alias::addAlias("trcssccr0", "s2_1_c1_c0_2");
-   Alias::addAlias("s3_5_c1_c0_2", "cpacr_el12");
-   Alias::addAlias("trcsscsr0", "s2_1_c1_c8_2");
-   Alias::addAlias("trcacatr0", "s2_1_c2_c0_2");
-   Alias::addAlias("s3_5_c2_c0_2", "tcr_el12");
-   Alias::addAlias("trccidcctlr0", "s2_1_c3_c0_2");
-   Alias::addAlias("trcvmidcctlr0", "s2_1_c3_c2_2");
-   Alias::addAlias("s3_0_c5_c4_2", "erxstatus_el1");
-   Alias::addAlias("s3_0_c9_c9_2", "pmsicr_el1");
-   Alias::addAlias("ich_misr_el2", "s3_4_c12_c11_2");
-   Alias::addAlias("s3_3_c14_c0_2", "cntvct_el0");
-   Alias::addAlias("s3_5_c14_c2_2", "cntp_cval_el02");
-   Alias::addAlias("s3_4_c14_c3_2", "cnthv_cval_el2");
-   Alias::addAlias("s3_5_c14_c3_2", "cntv_cval_el02");
-   Alias::addAlias("trcvictlr", "s2_1_c0_c0_2");
-   Alias::addAlias("trcviiectlr", "s2_1_c0_c1_2");
-   Alias::addAlias("s3_0_c0_c1_2", "id_dfr0_el1");
-   Alias::addAlias("trcvissctlr", "s2_1_c0_c2_2");
-   Alias::addAlias("trcvipcssctlr", "s2_1_c0_c3_2");
-   Alias::addAlias("trcvdctlr", "s2_1_c0_c8_2");
-   Alias::addAlias("trcvdsacctlr", "s2_1_c0_c9_2");
-   Alias::addAlias("trcvdarcctlr", "s2_1_c0_c10_2");
-   Alias::addAlias("trcssccr0", "s2_1_c1_c0_2");
-   Alias::addAlias("s3_5_c1_c0_2", "cpacr_el12");
-   Alias::addAlias("trcsscsr0", "s2_1_c1_c8_2");
-   Alias::addAlias("trcacatr0", "s2_1_c2_c0_2");
-   Alias::addAlias("s3_5_c2_c0_2", "tcr_el12");
-   Alias::addAlias("trccidcctlr0", "s2_1_c3_c0_2");
-   Alias::addAlias("trcvmidcctlr0", "s2_1_c3_c2_2");
-   Alias::addAlias("s3_0_c5_c4_2", "erxstatus_el1");
-   Alias::addAlias("s3_0_c9_c9_2", "pmsicr_el1");
-   Alias::addAlias("ich_misr_el2", "s3_4_c12_c11_2");
-   Alias::addAlias("s3_3_c14_c0_2", "cntvct_el0");
-   Alias::addAlias("s3_5_c14_c2_2", "cntp_cval_el02");
-   Alias::addAlias("s3_4_c14_c3_2", "cnthv_cval_el2");
-   Alias::addAlias("s3_5_c14_c3_2", "cntv_cval_el02");
-   Alias::addAlias("s3_0_c0_c1_3", "id_afr0_el1");
-   Alias::addAlias("trcsspcicr0", "s2_1_c1_c0_3");
-   Alias::addAlias("s3_0_c4_c2_3", "pan");
-   Alias::addAlias("s3_4_c5_c2_3", "vsesr_el2");
-   Alias::addAlias("s3_0_c5_c4_3", "erxaddr_el1");
-   Alias::addAlias("s3_0_c9_c9_3", "pmsirr_el1");
-   Alias::addAlias("s3_0_c9_c10_3", "pmbsr_el1");
-   Alias::addAlias("icc_bpr0_el1", "s3_0_c12_c8_3");
-   Alias::addAlias("s3_0_c0_c1_3", "id_afr0_el1");
-   Alias::addAlias("trcsspcicr0", "s2_1_c1_c0_3");
-   Alias::addAlias("s0_0_c4_c0_3", "uao");
-   Alias::addAlias("s3_0_c4_c2_3", "pan");
-   Alias::addAlias("s3_4_c5_c2_3", "vsesr_el2");
-   Alias::addAlias("s3_0_c5_c4_3", "erxaddr_el1");
-   Alias::addAlias("s3_0_c9_c9_3", "pmsirr_el1");
-   Alias::addAlias("s3_0_c9_c10_3", "pmbsr_el1");
-   Alias::addAlias("icc_bpr0_el1", "s3_0_c12_c8_3");
-   Alias::addAlias("trcseqevr0", "s2_1_c0_c0_4");
-   Alias::addAlias("s3_0_c0_c1_4", "id_mmfr0_el1");
-   Alias::addAlias("s3_0_c0_c5_4", "id_aa64afr0_el1");
-   Alias::addAlias("trcseqrstevr", "s2_1_c0_c6_4");
-   Alias::addAlias("trcseqstr", "s2_1_c0_c7_4");
-   Alias::addAlias("trcextinselr", "s2_1_c0_c8_4");
-   Alias::addAlias("trcoslar", "s2_1_c1_c0_4");
-   Alias::addAlias("s2_0_c1_c1_4", "oslsr_el1");
-   Alias::addAlias("trcpdcr", "s2_1_c1_c4_4");
-   Alias::addAlias("trcdvcvr0", "s2_1_c2_c0_4");
-   Alias::addAlias("s3_0_c4_c2_4", "uao");
-   Alias::addAlias("trcitctrl", "s2_1_c7_c0_4");
-   Alias::addAlias("s3_0_c9_c9_4", "pmsfcr_el1");
-   Alias::addAlias("ich_vseir_el2", "s3_4_c12_c9_4");
-   Alias::addAlias("icc_ctlr_el1", "s3_0_c12_c12_4");
-   Alias::addAlias("trcseqevr0", "s2_1_c0_c0_4");
-   Alias::addAlias("s3_0_c0_c1_4", "id_mmfr0_el1");
-   Alias::addAlias("s3_0_c0_c5_4", "id_aa64afr0_el1");
-   Alias::addAlias("trcseqrstevr", "s2_1_c0_c6_4");
-   Alias::addAlias("trcseqstr", "s2_1_c0_c7_4");
-   Alias::addAlias("trcextinselr", "s2_1_c0_c8_4");
-   Alias::addAlias("trcoslar", "s2_1_c1_c0_4");
-   Alias::addAlias("s2_0_c1_c1_4", "oslsr_el1");
-   Alias::addAlias("trcpdcr", "s2_1_c1_c4_4");
-   Alias::addAlias("trcdvcvr0", "s2_1_c2_c0_4");
-   Alias::addAlias("s0_0_c4_c0_4", "pan");
-   Alias::addAlias("s3_0_c4_c2_4", "uao");
-   Alias::addAlias("trcitctrl", "s2_1_c7_c0_4");
-   Alias::addAlias("s3_0_c9_c9_4", "pmsfcr_el1");
-   Alias::addAlias("ich_vseir_el2", "s3_4_c12_c9_4");
-   Alias::addAlias("icc_ctlr_el1", "s3_0_c12_c12_4");
-   Alias::addAlias("trccntrldvr0", "s2_1_c0_c0_5");
-   Alias::addAlias("s3_0_c0_c0_5", "mpidr_el1");
-   Alias::addAlias("trccntctlr0", "s2_1_c0_c4_5");
-   Alias::addAlias("trccntvr0", "s2_1_c0_c8_5");
-   Alias::addAlias("s3_0_c9_c9_5", "pmsevfr_el1");
-   Alias::addAlias("icc_sre_el2", "s3_4_c12_c9_5");
-   Alias::addAlias("icc_sgi1r_el1", "s3_0_c12_c11_5");
-   Alias::addAlias("trccntrldvr0", "s2_1_c0_c0_5");
-   Alias::addAlias("s3_0_c0_c0_5", "mpidr_el1");
-   Alias::addAlias("trccntctlr0", "s2_1_c0_c4_5");
-   Alias::addAlias("trccntvr0", "s2_1_c0_c8_5");
-   Alias::addAlias("spsel, 0x2", "msr s0_0_c4_c2_5");
-   Alias::addAlias("s3_0_c9_c9_5", "pmsevfr_el1");
-   Alias::addAlias("icc_sre_el2", "s3_4_c12_c9_5");
-   Alias::addAlias("icc_sgi1r_el1", "s3_0_c12_c11_5");
-   Alias::addAlias("s3_0_c0_c0_6", "revidr_el1");
-   Alias::addAlias("trcdvcmr0", "s2_1_c2_c0_6");
-   Alias::addAlias("trcclaimset", "s2_1_c7_c8_6");
-   Alias::addAlias("trcclaimclr", "s2_1_c7_c9_6");
-   Alias::addAlias("trclar", "s2_1_c7_c12_6");
-   Alias::addAlias("s2_0_c7_c14_6", "dbgauthstatus_el1");
-   Alias::addAlias("s3_0_c9_c9_6", "pmslatfr_el1");
-   Alias::addAlias("s3_3_c9_c12_6", "pmceid0_el0");
-   Alias::addAlias("icc_asgi1r_el1", "s3_0_c12_c11_6");
-   Alias::addAlias("icc_igrpen0_el1", "s3_0_c12_c12_6");
-   Alias::addAlias("s3_0_c0_c0_6", "revidr_el1");
-   Alias::addAlias("trcdvcmr0", "s2_1_c2_c0_6");
-   Alias::addAlias("trcclaimset", "s2_1_c7_c8_6");
-   Alias::addAlias("trcclaimclr", "s2_1_c7_c9_6");
-   Alias::addAlias("trclar", "s2_1_c7_c12_6");
-   Alias::addAlias("s2_0_c7_c14_6", "dbgauthstatus_el1");
-   Alias::addAlias("s3_0_c9_c9_6", "pmslatfr_el1");
-   Alias::addAlias("s3_3_c9_c12_6", "pmceid0_el0");
-   Alias::addAlias("icc_asgi1r_el1", "s3_0_c12_c11_6");
-   Alias::addAlias("icc_igrpen0_el1", "s3_0_c12_c12_6");
-   Alias::addAlias("trcimspec0", "s2_1_c0_c0_7");
-   Alias::addAlias("s3_1_c0_c0_7", "aidr_el1");
-   Alias::addAlias("s3_3_c0_c0_7", "dczid_el0");
-   Alias::addAlias("s3_0_c9_c9_7", "pmsidr_el1");
-   Alias::addAlias("s3_0_c9_c10_7", "pmbidr_el1");
-   Alias::addAlias("ich_vmcr_el2", "s3_4_c12_c11_7");
-   Alias::addAlias("trcimspec0", "s2_1_c0_c0_7");
-   Alias::addAlias("s3_1_c0_c0_7", "aidr_el1");
-   Alias::addAlias("s3_3_c0_c0_7", "dczid_el0");
-   Alias::addAlias("s3_0_c9_c9_7", "pmsidr_el1");
-   Alias::addAlias("s3_0_c9_c10_7", "pmbidr_el1");
-   Alias::addAlias("ich_vmcr_el2", "s3_4_c12_c11_7");
+
+   // Get a pointer in the buffer to move around and analyze bytes.
+   char* cur = buf;
+
+   // Go until we find the end of the string (which includes the immediate).
+   while(*cur) {
+      cur++;
+   }
+   cur--;
+
+   // Walk back to the decimal point, removing zeroes along the way.
+   bool inZeroes = true;
+   while (cur > buf && *cur != '.') {
+      if (*cur == '0' && inZeroes) {
+         *cur = 0;
+      } else {
+         inZeroes = false;
+      }
+      cur--;
+   }
+
+   // If we read only zeroes all the way to the decimal point, remove the
+   // decimal point as well.
+   if (inZeroes) {
+      *cur = 0;
+   }
+
+   // We hit the beginning of the buffer without a decimal point, so stop.
+   if (cur == buf) {
+      return;
+   }
+
+   // We need to see if the immediate value had a leading zero too, so back up.
+   cur--;
+
+   // If there isn't a zero, we know we're done already.
+   if (*cur != '0') {
+      return;
+   }
+   
+   if (*(cur - 1) == ' ' || (*(cur - 1) == '-' && *(cur - 2) == ' ')) {
+
+      while (*(cur + 1)) {
+         *cur = *(cur + 1);
+         cur++;
+      }
+
+      *cur = 0;
+   }
+
+}
+
+
+char negHex(char h) {
+   
+   if (h >= '0' && h <= '5') {
+      return 'f' + '0' - h;
+   }
+
+   if (h >= '6' && h <= '9') {
+      return '9' + '6' - h;
+   }
+
+   if (h >= 'a' && h <= 'f') {
+      return 'f' + '0' - h;
+   }
+
+   return h;
+}
+
+
+void aliasMovn(char* buf, int bufLen) {
+   if (strncmp(buf, "movn", 4)) {
+      return;
+   }
+
+   char* tmp = (char*)malloc(bufLen);
+   assert(tmp != NULL);
+
+   // We have a movz instruction, so delete the z.
+   char* cur = buf;
+   char* place = tmp;
+
+   for(int i = 0; i < 3; i++) {
+      *place = *cur;
+      place++;
+      cur++;
+   }
+
+   cur++;
+   bool wide = *(cur + 1) == 'x';
+   
+   // Copy until the immediate field.
+   while(*cur && *(cur - 4) != ',') {
+      *place = *cur;
+      place++;
+      cur++;
+   }
+
+   int nDigits = 0;
+   
+   while (*cur && *cur != ',') {
+      nDigits++;
+      cur++;
+   }
+   
+   int nFs = 0;
+   if (*cur != 0) {
+      nFs = 4 * (*(cur + 8) - '0');
+   }
+   cur -= nDigits;
+
+   // If we have all 4 digits but the first doesn't have a proper sign, put it
+   // there.
+   if ((nDigits < 4 || (!wide && nFs == 0) || (wide && nFs < 12)) && 
+         negHex(*cur) < '8' && negHex(*cur) >= '0') {
+      *place = 'f';
+      place++;
+   }
+
+   // Copy over the negated hex digits from the string.
+   for (int i = 0; i < nDigits; i++) {
+      *place = negHex(*cur);
+      place++;
+      cur++;
+   }
+
+   for (int i = 0; i < nFs; i++) {
+      *place = 'f';
+      place++;
+   }
+   *place = 0;
+
+   strcpy(buf, tmp);
+   free(tmp);
+}
+
+void aliasMovz(char* buf, int bufLen) {
+   if (strncmp(buf, "movz", 4)) {
+      return;
+   }
+
+   // We have a movz instruction, so delete the z.
+   char* cur = buf + 4;
+   int nCommas = 0;
+   while (*cur) {
+      if (*cur == ',') {
+         nCommas++;
+      }
+      *(cur - 1) = *cur;
+      cur++;
+   }
+   cur--;
+   *cur = 0;
+
+   // If we only ran into 1 comma, there wasn't a shift left term, so we are
+   // done already.
+   if (nCommas == 1) {
+      return;
+   }
+
+   cur -= 2;
+   int nZeroes = 4 * (*cur - '0');
+   
+   while (*cur != ',') {
+      cur--;
+   }
+
+   for (int i = 0; i < nZeroes; i++) {
+      *cur = '0';
+      cur++;
+   }
+   *cur = 0;
 }
