+++ fleece/decoders/capstone_aarch64.C
@@ -18,8 +18,61 @@
  * along with this software; if not, see www.gnu.org/licenses
 */
 
+#include "aarch64_common.h"
 #include "Normalization.h"
 #include "capstone/capstone.h"
+
+#include <iostream>
+
+void makeIndexesDecimal(char* buf, int bufLen) {
+
+    char tmp[bufLen];
+    char* tmpStart = &tmp[0];
+    char* cur = buf;
+    while (*cur) {
+        if (/**cur == '}' &&*/ *(cur + 1) && *(cur + 1) == '[' && 
+                           *(cur + 2) && *(cur + 2) == '0' && 
+                           *(cur + 3) && *(cur + 3) == 'x') {
+            cur += 2;
+            char* place = cur;
+            char* digitsEnd = NULL;
+            long long int val = strtoll(cur, &digitsEnd, 16);
+            strcpy(tmpStart, digitsEnd);
+            place += snprintf(place, bufLen - (place - buf), "%lld", val);
+            strncpy(place, tmpStart, bufLen - (place - buf));
+        }
+        cur++;
+    }
+}
+
+void makeHexConstantsPositive(char* buf, int bufLen) {
+
+    char tmp[bufLen];
+    char* tmpStart = &tmp[0];
+
+    char* place = buf;
+    char* cur = buf;
+    while (*cur) {
+        if (*cur == '-' && *(cur + 1) && *(cur + 1) == '0' && 
+                           *(cur + 2) && *(cur + 2) == 'x') {
+            
+            char* digitsEnd = NULL;
+            long long int val = strtoll(cur, &digitsEnd, 16);
+            strcpy(tmpStart, digitsEnd);
+            *place = '0';
+            place++;
+            *place = 'x';
+            place++;
+            int nDigits = snprintf(place, bufLen - (place - buf), "%llx", val);
+            place += nDigits;
+            cur = tmpStart;
+        }
+        *place = *cur;
+        place++;
+        cur++;
+    }
+    *place = '\0';
+}
 
 int capstone_aarch64_decode(char* inst, int nBytes, char* buf, int bufLen) {
 
@@ -31,7 +84,7 @@ int capstone_aarch64_decode(char* inst, int nBytes, char* buf, int bufLen) {
       return -1;
    }
 
-   int nInsns = cs_disasm(handle, (uint8_t*)inst, nBytes, 0x1000, 0, &insn);
+   int nInsns = cs_disasm(handle, (uint8_t*)inst, nBytes, 0, 0, &insn);
    
    if (nInsns < 1) {
       return -1;
@@ -45,5 +98,14 @@ int capstone_aarch64_decode(char* inst, int nBytes, char* buf, int bufLen) {
 }
 
 void capstone_aarch64_norm(char* buf, int bufLen) {
-
+    removePounds(buf, bufLen);
+    removeExtraZeroesFromFmovImm(buf, bufLen);
+    decToHexConstants(buf, bufLen);
+    makeIndexesDecimal(buf, bufLen);
+    makeHexConstantsPositive(buf, bufLen);
+    aliasMovz(buf, bufLen);
+    aliasMovn(buf, bufLen);
+    trimHexFs(buf, bufLen);
+    trimHexZeroes(buf, bufLen);
+    removeADRPZeroes(buf, bufLen);
 }
